Use accessor history for requested frame in affine transform

+++ tiamat/readers/processing.py
@@ -170,7 +170,7 @@ def access_image(image: np.ndarray, accessor: ImageAccessor, image_scale: tuple[
 
     # Loop over all dimensions to create request
     request_slices = [slice(None)] * len(image.shape)
-    access_shape = np.ones((len(image.shape)))
+    access_shape = np.ones((len(image.shape)), dtype=np.int64)
     for dim, coord in zip(ch_dims + image_dims, access_ch_dims + access_image_dims):
         max_c = image.shape[dim]
         c_from, c_to = coord
